+++ CMakeLists.txt
@@ -7,10 +7,7 @@ set(CMAKE_CXX_STANDARD 17)
 add_subdirectory(search_lib)
 
 
-target_include_directories(${PROJECT_NAME} PUBLIC ${SEARCH_LIB_INCLUDE_DIRS})
-target_link_libraries(${PROJECT_NAME} PRIVATE ${SEARCH_LIB_LIBRARIES})
-
-add_subdirectory(tests)
+# add_subdirectory(tests)
 
 
 set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -std=c++17 -Wall")
@@ -36,6 +33,11 @@ add_executable(tgbot main.cpp)
 
 target_link_libraries(tgbot PRIVATE TGBOT_LIB)
 target_link_libraries(tgbot /usr/local/lib/libTgBot.a ${CMAKE_THREAD_LIBS_INIT} ${OPENSSL_LIBRARIES} ${Boost_LIBRARIES} ${CURL_LIBRARIES})
+
+
+target_include_directories(${PROJECT_NAME} PUBLIC ${SEARCH_LIB_INCLUDE_DIRS})
+target_link_libraries(${PROJECT_NAME} PRIVATE ${SEARCH_LIB_LIBRARIES})
+
 
 find_package(GTest REQUIRED)
 # find_package(GMock REQUIRED)
