+++ lib/tg_bot.cpp
@@ -8,6 +8,10 @@
 #include <iostream>
 #include <string>
 
+#include "../search_lib/search.h"
+#include "../search_lib/database.h"
+#include "../search_lib/database.h"
+
 // #include "tg_bot.h"
 
 using TgBot::Bot;
@@ -42,10 +46,183 @@ inline constexpr auto hash_djb2a(const std::string_view sv) {
 inline constexpr auto operator"" _sh(const char *str, size_t len) {
     return hash_djb2a(std::string_view{ str, len });
 }
-struct Route
-{
-    std::vector<std::string> images;
+// struct Route
+// {
+//     std::vector<std::string> images;
+// };
+
+std::vector <BasePoint> fillTestBasePoints1() {
+    BasePoint a(1), b(2), c(3), d(4);
+    // base
+    {
+        Edge edge(1, 2, 2, "1->2.txt");
+        a.AddEdge(edge);
+    }
+    {
+        Edge edge(1, 4, 1, "1->4.txt");
+        a.AddEdge(edge);
+    }
+    // infr
+    {
+        Edge edge(1, 9, 5, "1->9.txt");
+        a.AddEdge(edge);
+    }
+    {
+        Edge edge(1, 10, 2, "1->10.txt");
+        a.AddEdge(edge);
+    }
+    {
+        Edge edge(1, 11, 2, "1->11.txt");
+        a.AddEdge(edge);
+    }
+    {
+        Edge edge(1, 102, 1, "1->canteen.txt");
+        a.AddEdge(edge);
+    }
+
+    // base
+    {
+        Edge edge(2, 1, 2, "2->1.txt");
+        b.AddEdge(edge);
+    }
+    {
+        Edge edge(2, 3, 5, "2->3.txt");
+        b.AddEdge(edge);
+    }
+    // infr
+    {
+        Edge edge(2, 10, 0, "2->10.txt");
+        b.AddEdge(edge);
+    }
+    {
+        Edge edge(2, 11, 1, "2->11.txt");
+        b.AddEdge(edge);
+    }
+    {
+        Edge edge(2, 12, 2, "2->12.txt");
+        b.AddEdge(edge);
+    }
+    {
+        Edge edge(2, 13, 3, "2->13.txt");
+        b.AddEdge(edge);
+    }
+
+    // base
+    {
+        Edge edge(3, 2, 5, "3->2.txt");
+        c.AddEdge(edge);
+    }
+    {
+        Edge edge(3, 4, 2, "3->4.txt");
+        c.AddEdge(edge);
+    }
+    // infr
+    {
+        Edge edge(3, 12, 1, "3->12.txt");
+        c.AddEdge(edge);
+    }
+    {
+        Edge edge(3, 13, 1, "3->13.txt");
+        c.AddEdge(edge);
+    }
+
+    // base
+    {
+        Edge edge(d.GetId(), 1, 1, "4->1.txt");
+        d.AddEdge(edge);
+    }
+    {
+        Edge edge(d.GetId(), 3, 2, "4->3.txt");
+        d.AddEdge(edge);
+    }
+    // infr
+    {
+        Edge edge(d.GetId(), 103, 1, "4->canteen.txt");
+        d.AddEdge(edge);
+    }
+
+    std::vector <BasePoint> graf = {a, b, c, d};
+    return graf;
+}
+
+std::vector <Infrastructure> fillInfrPoints1() {
+    Infrastructure i9(9), i10(10), i11(11), i12(12), i13(13), i102(102), i103(103);
+
+    i9.AddName("399u");
+    {
+        Edge edge(i9.GetId(), 1, 5, "9->1.txt");
+        i9.AddEdge(edge);
+    }
+
+    i10.AddName("400u");
+    {
+        Edge edge(i10.GetId(), 1, 2, "10->1.txt");
+        i10.AddEdge(edge);
+    }
+    {
+        Edge edge(i10.GetId(), 2, 0, "10->2.txt");
+        i10.AddEdge(edge);
+    }
+
+    i11.AddName("401u");
+    {
+        Edge edge(i11.GetId(), 1, 2, "11->1.txt");
+        i11.AddEdge(edge);
+    }
+    {
+        Edge edge(i11.GetId(), 2, 1, "11->2.txt");
+        i11.AddEdge(edge);
+    }
+
+    i12.AddName("402u");
+    {
+        Edge edge(i12.GetId(), 2, 2, "12->2.txt");
+        i12.AddEdge(edge);
+    }
+    {
+        Edge edge(i12.GetId(), 3, 1, "12->3.txt");
+        i12.AddEdge(edge);
+    }
+
+    i13.AddName("403u");
+    {
+        Edge edge(i13.GetId(), 2, 2, "13->2.txt");
+        i13.AddEdge(edge);
+    }
+    {
+        Edge edge(i13.GetId(), 3, 1, "13->3.txt");
+        i13.AddEdge(edge);
+    }
+
+    i102.AddName("canteen");
+    i102.AddName("canteen2");
+    {
+        Edge edge(i102.GetId(), 1, 1, "canteen->1.txt");
+        i102.AddEdge(edge);
+    }
+
+    i103.AddName("canteen");
+    i103.AddName("canteen3");
+    i103.AddName("canteen_near_physics");
+    {
+        Edge edge(i103.GetId(), 4, 1, "canteen->4.txt");
+        i103.AddEdge(edge);
+    }
+
+    std::vector <Infrastructure> graf = {i9, i10, i11, i12, i13, i102, i103};
+    return graf;
+}
+
+class DataBaseTest : public DataBase {
+public:
+    std::vector <BasePoint> getBasePoints() override {
+        return fillTestBasePoints1();
+    }
+    std::vector <Infrastructure> getInfrastructurePoints() override {
+        return fillInfrPoints1();
+    }
 };
+
 
 struct Positions
 {
@@ -55,42 +232,42 @@ struct Positions
     int currentPointer; 
 };
 
-class DataBase
-{
+// class DataBase
+// {
 
-};
+// };
 
-class DataBaseBMSTU : public DataBase
-{
+// class DataBaseBMSTU : public DataBase
+// {
 
-};
+// };
 
 
-class Search
-{
-public:
-    Search (DataBase *db) : db(db) {}
-    Route *FindRoute(const Positions &pos)
-    {
-        Route *route = new Route; 
-        route->images.push_back("mem.jpg"); 
-        route->images.push_back("apple.jpg"); 
-        // route->images.push_back("video.mp4"); 
-        return route;
-    }
-    bool HavePoint(std::string point)
-    {
-        return true;
-    }
-private:
-    DataBase *db;
-};
+// class Search
+// {
+// public:
+//     Search (DataBase *db) : db(db) {}
+//     Route *FindRoute(const Positions &pos)
+//     {
+//         Route *route = new Route; 
+//         route->images.push_back("mem.jpg"); 
+//         route->images.push_back("apple.jpg"); 
+//         // route->images.push_back("video.mp4"); 
+//         return route;
+//     }
+//     bool HavePoint(std::string point)
+//     {
+//         return true;
+//     }
+// private:
+//     DataBase *db;
+// };
 
 class IModel
 {
 public:
     IModel(DataBase *db) : search(db) {};
-    virtual Route *FindRouteModel(const Positions &pos) = 0;
+    virtual std::vector<std::string> FindRouteModel(const Positions &pos) = 0;
     virtual bool isValid(std::string point) = 0;
 protected:
     Search search;
@@ -100,13 +277,23 @@ class Model : public IModel
 {
 public:
     Model(DataBase *db) : IModel(db) {}
-    Route *FindRouteModel(const Positions &pos) 
+    std::vector<std::string> FindRouteModel(const Positions &pos) 
     {
-        Route *route = search.FindRoute(pos);
-        return route;
+        std::vector<Point*> points = search.GetByName(pos.start);
+        if (points.size())
+        {
+            unsigned int id = points[0]->GetId();
+            Route route = search.FindRoute(id, pos.end);
+            std::vector<std::string> vecPath = route.GetLinks();
+            return vecPath;
+        }
+        throw std::runtime_error("error");
+        // Route *route = search.FindRoute(pos);
+        // return route;
     }
     bool isValid(std::string point)
     {
+        std::cout << point << " " << search.HavePoint(point) << std::endl;
         return search.HavePoint(point);
     }
 };
@@ -312,7 +499,6 @@ private:
         {
             case "dining"_sh:
             case "audit"_sh:
-            case "dining2"_sh:
             case "toilet"_sh:
             {
                 Message outputMsg("Выберите: " + std::string(j.at(CATEG).at(msg.GetText()).at(TEXT)), msg.GetChatId());
@@ -403,9 +589,12 @@ private:
             {
                 if (isReady())
                 {
-                    Route *route = model->FindRouteModel(GetPos());
-                    Message videoMsg(route->images, msg.GetChatId());
-                    if (IView *view = FindView(VIDEO))
+                    // Route *route = model->FindRouteModel(GetPos());
+                    std::vector<std::string> vecPath = model->FindRouteModel(GetPos());
+                    std::string str; 
+                    for (auto &elem : vecPath) str += elem; 
+                    Message videoMsg(str, msg.GetChatId());
+                    if (IView *view = FindView(TEXT))
                     {
                         view->SendMessage(videoMsg);
                     }
