+++ search_lib/database.h
@@ -1,4 +1,7 @@
-#pragma once
+
+#ifndef DATABASE_H
+#define DATABASE_H
+
 #include <iostream>
 #include <vector>
 #include "points.h"
@@ -8,4 +11,7 @@ public:
     virtual std::vector <BasePoint> getBasePoints() = 0;
     virtual std::vector <Infrastructure> getInfrastructurePoints() = 0;
 };
+
 // class DataBase;
+
+#endif
