+++ tests/CmakeLists.txt
@@ -1,12 +1,12 @@
-cmake_minimum_required(VERSION 3.14)
-project(test_search)
+# cmake_minimum_required(VERSION 3.14)
+# project(test_search)
 
-set(CMAKE_CXX_STANDARD 17)
+# set(CMAKE_CXX_STANDARD 17)
 
-file(GLOB SOURCES *.cpp)
+# file(GLOB SOURCES *.cpp)
 
-find_package(GTest REQUIRED)
+# find_package(GTest REQUIRED)
 
-add_executable(${PROJECT_NAME} ${SOURCES})
-target_include_directories(${PROJECT_NAME} PUBLIC ${SEARCH_LIB_INCLUDE_DIRS})
-target_link_libraries(${PROJECT_NAME} ${SEARCH_LIB_LIBRARIES} GTest::gtest_main)
+# add_executable(${PROJECT_NAME} ${SOURCES})
+# target_include_directories(${PROJECT_NAME} PUBLIC ${SEARCH_LIB_INCLUDE_DIRS})
+# target_link_libraries(${PROJECT_NAME} ${SEARCH_LIB_LIBRARIES} GTest::gtest_main)
