+++ tests/gtests.cpp
@@ -1,12 +1,12 @@
 #include <iostream>
-#include <sys/sysinfo.h>
+// #include <sys/sysinfo.h>
 #include <gtest/gtest.h>
-#include "gmock/gmock.h"
+// #include "gmock/gmock.h"
 
-using ::testing::AtLeast;
-using ::testing::DoAll;
-using ::testing::Return;
-using ::testing::SetArgReferee;
+// using ::testing::AtLeast;
+// using ::testing::DoAll;
+// using ::testing::Return;
+// using ::testing::SetArgReferee;
 
 #include "tg_bot.h"
